fat16: separate FAT16 file system to fs/fat16
