+++ fs/fat16/dump.c
@@ -1,129 +1,6 @@
 #include <fs/fat16.h>
-#include <ide.h>
 #include <kernel.h>
 #include <lib/string.h>
-#include <mm.h>
-
-void fat16_dump_mbr(FAT16_MBR *mbr);
-void fat16_dump_bpb(FAT16_BPB *bpb);
-void fat16_dump_dir(FAT16_DIR_ENTRY *dir, uint32 entries);
-void fat16_read_rootdir(FAT16_DIR_ENTRY *rootdir);
-
-static FAT16 fat16;
-
-void fat16_init() {
-    /*
-        Read MBR
-    */
-    FAT16_MBR *mbr = (FAT16_MBR *)kmalloc(sizeof(FAT16_MBR));
-    ide_read(0, mbr);
-#ifdef __DEBUG__
-    fat16_dump_mbr(mbr);
-#endif
-    // TODO: Support multiple partitions
-    if (mbr->bootsig != 0xAA55 || mbr->table[0].boot != 0x80 ||
-        (mbr->table[0].type != 0x04 && mbr->table[0].type != 0x06)) {
-        panic("FAT16: Invalid MBR\n");
-    }
-    fat16.bpb_entry = mbr->table[0].bpb_begin;
-    kprintf("[fs] FAT16 Partition0 found\n");
-
-    /*
-        Read BPB
-    */
-    FAT16_BPB *bpb = (FAT16_BPB *)kmalloc(sizeof(FAT16_BPB));
-    ide_read(fat16.bpb_entry, bpb);
-#ifdef __DEBUG__
-    fat16_dump_bpb(bpb);
-#endif
-    if (bpb->bytes_per_sector != SECTOR_SIZE || bpb->fats != 2 ||
-        (bpb->signature != 0x28 && bpb->signature != 0x29)) {
-        panic("FAT16: Invalid BPB\n");
-    }
-    fat16.sector_size = bpb->bytes_per_sector;
-    fat16.cluster_size = bpb->bytes_per_sector * bpb->sectors_per_cluster;
-    fat16.fat_size = bpb->bytes_per_sector * bpb->sectors_per_fat;
-    fat16.partition0_size = mbr->table[0].sectors * bpb->bytes_per_sector;
-    fat16.sector_per_cluster = bpb->sectors_per_cluster;
-    fat16.sector_per_fat = bpb->sectors_per_fat;
-    fat16.sector_per_partition0 = bpb->sectors;
-    fat16.fat_entry = fat16.bpb_entry + bpb->reserved_sectors;
-    fat16.root_entries = bpb->root_entries;
-    fat16.rootdir_entry = fat16.fat_entry + bpb->sectors_per_fat * bpb->fats;
-    fat16.data_entry = fat16.rootdir_entry + sizeof(FAT16_DIR_ENTRY) *
-                                                 bpb->root_entries /
-                                                 bpb->bytes_per_sector;
-
-    kprintf("[fs] Partition Size: %d KiB\n", fat16.partition0_size / KiB(1));
-    kprintf("[fs] Cluster Size: %d KiB\n", fat16.cluster_size / KiB(1));
-    kprintf("[fs] FAT Size: %d KiB\n", fat16.fat_size / KiB(1));
-    kprintf("[fs] Volume Label: %s\n", bpb->volume_label);
-
-    kmfree(mbr);
-    kmfree(bpb);
-    return;
-}
-
-uint16 fat16_find_cluster(char *filename, uint16 dir_cluster) {
-    FAT16_DIR_ENTRY *dir;
-    uint16 entries;
-    if (dir_cluster == 0) {
-        entries = fat16.root_entries;
-        dir = (FAT16_DIR_ENTRY *)kmalloc(sizeof(FAT16_DIR_ENTRY) * entries);
-        fat16_read_rootdir(dir);
-    } else {
-        entries = fat16.cluster_size / sizeof(FAT16_DIR_ENTRY);
-        dir = (FAT16_DIR_ENTRY *)kmalloc(fat16.cluster_size);
-        fat16_read_cluster(dir_cluster, (char *)dir);
-    }
-
-    for (int i = 0; i < entries; i++) {
-        if (dir[i].filename[0] == 0x00) {
-            continue;
-        }
-        if (dir[i].attr == 0x0F) {
-            continue;
-        }
-        char entry_filename[12];
-        for (int j = 0; j < 11; j++) {
-            entry_filename[j] = dir[i].filename[j];
-            if (entry_filename[j] == ' ') {
-                entry_filename[j] = '\0';
-            }
-        }
-        entry_filename[11] = '\0';
-        if (strcmp(filename, entry_filename) == 0) {
-            kmfree(dir);
-            return dir[i].cluster;
-        }
-    }
-
-    kmfree(dir);
-    return 0;
-}
-
-uint16 fat16_next_cluster(uint16 cluster) {
-    uint16 *fat = (uint16 *)kmalloc(fat16.sector_size);
-
-    // 256 = 512(sector size) / 2(FAT entry size)
-    uint16 fat_entries = fat16.sector_size / sizeof(uint16);
-    ide_read(fat16.fat_entry + (cluster / fat_entries), fat);
-    uint16 next = fat[cluster % fat_entries];
-
-    kmfree(fat);
-    return next;
-}
-
-void fat16_read_cluster(uint16 cluster, char *buf) {
-    ide_read_seq(fat16.data_entry + (cluster - 2) * fat16.sector_per_cluster,
-                 buf, fat16.sector_per_cluster);
-}
-
-void fat16_read_rootdir(FAT16_DIR_ENTRY *rootdir) {
-    ide_read_seq(
-        fat16.rootdir_entry, rootdir,
-        sizeof(FAT16_DIR_ENTRY) * fat16.root_entries / fat16.sector_size);
-}
 
 void fat16_dump_mbr(FAT16_MBR *mbr) {
     debugf("FAT16 MBR:\n");
