--- fs/init.c
@@ -1,5 +1,0 @@
-#include <fs.h>
-#include <fs/fat16.h>
-
-extern void fat16_init();
-void fs_init() { fat16_init(); }
