+++ include/fs/fat16.h
@@ -1,5 +1,4 @@
 #pragma once
-#include <fs.h>
 #include <types.h>
 
 /* references:
